feat: Implement decimal quantities and weight-based pricing
This commit introduces support for fractional quantities (e.g., kilograms/grams) across the sales workflow and refines the "Save Order" system to be more descriptive and user-centric.

### 1. New: Weight-Based Quantity Support
The system now supports decimal quantities, allowing for the sale of products by weight.
- **Decimal Quantities (`cart_item.dart`, `QuickSale.dart`):** Updated the `CartItem` and `QuickSaleItem` models to use `double` for quantities instead of `int`.
- **Weight Input Dialog (`saleall.dart`):** Introduced a specialized dialog for items with "kg" or "kilogram" units, allowing users to enter weight in either grams or kilograms.
- **Stock Validation:** Updated stock decrement logic and validation prompts to handle decimal values correctly while maintaining integrity for integer-based stock tracking.
- **UI Formatting:** Quantities in the cart and sales views are now dynamically formatted (e.g., showing "0.5" for half a kg while keeping "1" for unit-based items).

### 2. Enhanced: Saved Order Management
The "Save Order" functionality has been refactored to focus on identifiable order names rather than just customer contact details.
- **Named Orders (`common_widgets.dart`, `Saved.dart`):** Replaced the requirement for customer name/phone with a mandatory "Order Name" field, making it easier for staff to identify specific parked carts.
- **Improved UI Persistence (`NewSale.dart`, `saleall.dart`):**
    - The active saved order name is now displayed prominently in the action bar with a distinct orange bookmark icon.
    - Improved `CartService` integration ensures the `savedOrderId` is correctly tracked and updated when a loaded order is modified.
- **Backward Compatibility:** Existing saved orders without an explicit "orderName" now fall back to using the "customerName" field for display.

### 3. Sales Workflow Improvements
- **Cart Synchronization (`NewSale.dart`):** Enhanced state management when loading saved orders, ensuring items and order IDs are correctly synced across the "View All" and "Quick Sale" interfaces.
- **Dynamic Quantity Controls:** Updated the edit quantity dialogs to allow incrementing/decrementing by 0.1 for weights less than 1kg.
- **Cart Cleaning Logic:** Clearing the cart now correctly resets all saved order tracking states and metadata.

### 4. Bug Fixes & Refinement
- **Stock Tracking (`Bill.dart`, `QuickSale.dart`):** Fixed stock decrement logic to ensure compatibility between decimal cart quantities and integer stock values.
- **UI Consistency:** Standardized "Cancel" button casing to Title Case in various dialogs and refined the layout of the saved orders list.

+++ lib/Sales/nq.dart
@@ -292,7 +292,7 @@ class _NewQuotationPageState extends State<NewQuotationPage> with SingleTickerPr
                           onPressed: exceedsStock ? null : () {
                             final newName = nameController.text.trim();
                             final newPrice = double.tryParse(priceController.text.trim()) ?? item.price;
-                            final newQty = int.tryParse(qtyController.text.trim()) ?? 1;
+                            final newQty = double.tryParse(qtyController.text.trim()) ?? 1.0;
 
                             // Final stock validation
                             if (stockEnabled && newQty > availableStock) {
